Move chat related models to ChatBasic
