+++ Core/Sources/ChatService/ChatFunctionProvider.swift
@@ -1,3 +1,4 @@
+import ChatBasic
 import Foundation
 import OpenAIService
