+++ Tool/Package.swift
@@ -170,6 +170,16 @@ let package = Package(
             name: "SuggestionModelTests",
             dependencies: ["SuggestionModel"]
         ),
+        
+        .target(
+            name: "ChatBasic",
+            dependencies: [
+                "AIModel",
+                "Preferences",
+                "Keychain",
+                .product(name: "CodableWrappers", package: "CodableWrappers"),
+            ]
+        ),
 
         .target(name: "AXExtension"),
 
@@ -205,8 +215,10 @@ let package = Package(
             name: "BuiltinExtension",
             dependencies: [
                 "SuggestionModel",
+                "ChatBasic",
                 "Workspace",
                 "ChatTab",
+                "AIModel",
                 .product(name: "CopilotForXcodeKit", package: "CopilotForXcodeKit"),
             ]
         ),
@@ -288,6 +300,7 @@ let package = Package(
                 "OpenAIService",
                 "ObjectiveCExceptionHandling",
                 "USearchIndex",
+                "ChatBasic",
                 .product(name: "Parsing", package: "swift-parsing"),
                 .product(name: "SwiftSoup", package: "SwiftSoup"),
             ]
@@ -312,6 +325,7 @@ let package = Package(
             dependencies: [
                 "LanguageClient",
                 "SuggestionModel",
+                "ChatBasic",
                 "Logger",
                 "Preferences",
                 "Terminal",
@@ -354,6 +368,7 @@ let package = Package(
                 "Preferences",
                 "TokenEncoder",
                 "Keychain",
+                "BuiltinExtension",
                 .product(name: "JSONRPC", package: "JSONRPC"),
                 .product(name: "AsyncAlgorithms", package: "swift-async-algorithms"),
                 .product(name: "GoogleGenerativeAI", package: "generative-ai-swift"),
@@ -390,6 +405,7 @@ let package = Package(
             name: "ChatContextCollector",
             dependencies: [
                 "SuggestionModel",
+                "ChatBasic",
                 "OpenAIService",
             ]
         ),
