+++ Tool/Sources/ChatBasic/JSONSchema.swift
