+++ Tool/Sources/GitHubCopilotService/LanguageServer/GitHubCopilotRequest.swift
@@ -13,7 +13,6 @@ struct GitHubCopilotDoc: Codable {
     var relativePath: String
     var languageId: CodeLanguage
     var position: Position
-    /// Buffer version. Not sure what this is for, not sure how to get it
     var version: Int = 0
 }
