+++ Tool/Sources/LangChain/Chains/RelevantInformationExtractionChain.swift
@@ -1,3 +1,4 @@
+import ChatBasic
 import Foundation
 import OpenAIService
