+++ Tool/Sources/OpenAIService/Memory/AutoManagedChatGPTMemoryStrategy/AutoManagedChatGPTMemoryOpenAIStrategy.swift
@@ -1,3 +1,4 @@
+import ChatBasic
 import Foundation
 import Logger
 import TokenEncoder
